+++ src/repository/xtream_repository.rs
@@ -1,16 +1,23 @@
 use std::cell::Ref;
 use std::collections::{BTreeMap, HashMap};
 use std::fs;
-use std::fs::File;
+use std::fs::{File, OpenOptions};
 use std::io::{BufReader, BufWriter, Error, Read, Seek, SeekFrom, Write};
 use std::iter::FromIterator;
 use std::path::{Path, PathBuf};
+use log::error;
 use serde::Serialize;
 use serde_json::{json, Map, Value};
 use crate::model::config::{Config, ConfigTarget};
 use crate::model::model_m3u::{PlaylistGroup, PlaylistItemHeader, XtreamCluster};
 use crate::{create_m3u_filter_error_result, utils};
+use crate::api::api_model::AppState;
 use crate::m3u_filter_error::{M3uFilterError, M3uFilterErrorKind};
+use crate::model::api_proxy::UserCredentials;
+use crate::utils::{get_client_request};
+
+type IndexTree = BTreeMap<i32, (u32, u16)>;
+
 
 pub(crate) static COL_CAT_LIVE: &str = "cat_live";
 pub(crate) static COL_CAT_SERIES: &str = "cat_series";
@@ -34,6 +41,27 @@ const SERIES_STREAM_FIELDS: &[&str] = &[
     "stream_type", "title", "year", "youtube_trailer",
 ];
 
+
+pub(crate) fn get_xtream_storage_path(cfg: &Config, target_name: &str) -> Option<PathBuf> {
+    utils::get_file_path(&cfg.working_dir, Some(std::path::PathBuf::from(target_name.replace(' ', "_"))))
+}
+
+pub(crate) fn get_xtream_epg_file_path(path: &Path) -> PathBuf {
+    path.join("epg.xml")
+}
+
+fn get_collection_path(path: &Path, collection: &str) -> PathBuf {
+    path.join(format!("{}.json", collection))
+}
+
+fn get_info_collection_path(path: &Path, collection: &str) -> PathBuf {
+    path.join(format!("{}_info.db", collection))
+}
+
+fn get_info_idx_path(path: &Path, collection: &str) -> PathBuf {
+    path.join(format!("{}_info.idx", collection))
+}
+
 fn write_to_file<T>(file: &Path, value: &T) -> Result<(), Error>
     where
         T: ?Sized + Serialize {
@@ -50,68 +78,40 @@ fn write_to_file<T>(file: &Path, value: &T) -> Result<(), Error>
     }
 }
 
-fn get_collection_and_idx_path(path: &Path, cluster: &XtreamCluster) -> (PathBuf, PathBuf) {
+fn get_info_collection_and_idx_path(path: &Path, cluster: &XtreamCluster) -> (PathBuf, PathBuf) {
     let collection = match cluster {
         XtreamCluster::Live => COL_LIVE,
         XtreamCluster::Video => COL_VOD,
         XtreamCluster::Series => COL_SERIES,
     };
-    (get_collection_path(path, collection), get_idx_path(path, collection))
+    (get_info_collection_path(path, collection), get_info_idx_path(path, collection))
 }
 
-fn write_to_file_width_idx(path: &Path, values: &[(i32, Value)], cluster: &XtreamCluster) -> Result<(), Error> {
-    let (file, file_idx) = get_collection_and_idx_path(path, cluster);
-    match File::create(file) {
-        Ok(file) => {
-            let mut index = BTreeMap::<i32, (u32, u16)>::new();
-            let mut writer = BufWriter::new(file);
-            writer.write_all("[".as_bytes())?;
-            let mut offset = 1;
-            let value_cnt = values.len();
-            let mut value_idx = 0;
-            for (stream_id, data) in values {
-                let content = serde_json::to_string(data).unwrap();
-                let bytes = content.as_bytes();
-                let size = bytes.len();
-                index.insert(*stream_id, (offset as u32, size as u16));
-                offset += size;
-                let _ = writer.write_all(bytes);
-                value_idx += 1;
-                if value_idx < value_cnt {
-                    writer.write_all(",".as_bytes())?;
-                    offset += 1;
-                }
+fn write_xtream_info(app_state: &AppState, target_name: &str, stream_id: i32, cluster: &XtreamCluster, content: &String, index_tree: &mut IndexTree) -> Result<(), Error> {
+    if let Some(path) = get_xtream_storage_path(&app_state.config, target_name) {
+        let (col_path, idx_path) = get_info_collection_and_idx_path(&path, cluster);
+        let mut comp: Vec<u8> = Vec::new();
+        lzma_rs::lzma_compress(&mut BufReader::new(content.as_bytes()), &mut comp)?;
+        let size = comp.len();
+        let lock = app_state.shared_locks.get_lock(target_name);
+        let shared_lock = lock.write().unwrap();
+        match OpenOptions::new()
+            .create(true)
+            .write(true)
+            .append(true)
+            .open(col_path) {
+            Ok(mut file) => {
+                let offset = file.metadata().unwrap().len();
+                file.write_all(comp.as_slice())?;
+                file.flush()?;
+                index_tree.insert(stream_id, (offset as u32, size as u16));
+                write_index(&idx_path, index_tree)?;
+                drop(shared_lock);
             }
-            writer.write_all("]".as_bytes())?;
-            match writer.flush() {
-                Ok(_) => {
-                    let encoded: Vec<u8> = bincode::serialize(&index).unwrap();
-                    let _ = fs::write(file_idx, encoded);
-                    Ok(())
-                }
-                Err(e) => Err(e)
-            }
+            Err(err) => return Err(err)
         }
-        Err(e) => Err(e)
     }
-}
-
-
-pub(crate) fn get_xtream_storage_path(cfg: &Config, target_name: &str) -> Option<PathBuf> {
-    utils::get_file_path(&cfg.working_dir, Some(std::path::PathBuf::from(target_name.replace(' ', "_"))))
-}
-
-fn get_collection_path(path: &Path, collection: &str) -> PathBuf {
-    path.join(format!("{}.json", collection))
-}
-
-fn get_idx_path(path: &Path, collection: &str) -> PathBuf {
-    path.join(format!("{}.idx", collection))
-}
-
-
-pub(crate) fn get_xtream_epg_file_path(path: &Path) -> PathBuf {
-    path.join("epg.xml")
+    Ok(())
 }
 
 pub(crate) fn write_xtream_playlist(target: &ConfigTarget, cfg: &Config, playlist: &[PlaylistGroup]) -> Result<(), M3uFilterError> {
@@ -221,7 +221,7 @@ pub(crate) fn write_xtream_playlist(target: &ConfigTarget, cfg: &Config, playlis
                         XtreamCluster::Live => &mut live_col,
                         XtreamCluster::Series => &mut series_col,
                         XtreamCluster::Video => &mut vod_col,
-                    }.push((stream_id, Value::Object(document)));
+                    }.push(Value::Object(document));
                 }
             }
         }
@@ -230,22 +230,14 @@ pub(crate) fn write_xtream_playlist(target: &ConfigTarget, cfg: &Config, playlis
         for (col_path, data) in [
             (get_collection_path(&path, COL_CAT_LIVE), &cat_live_col),
             (get_collection_path(&path, COL_CAT_VOD), &cat_vod_col),
-            (get_collection_path(&path, COL_CAT_SERIES), &cat_series_col)] {
+            (get_collection_path(&path, COL_CAT_SERIES), &cat_series_col),
+            (get_collection_path(&path, COL_LIVE), &live_col),
+            (get_collection_path(&path, COL_VOD), &vod_col),
+            (get_collection_path(&path, COL_SERIES), &series_col)] {
             match write_to_file(&col_path, data) {
                 Ok(()) => {}
                 Err(err) => {
                     errors.push(format!("Persisting collection failed: {}: {}", &col_path.to_str().unwrap(), err));
-                }
-            }
-        }
-        for (data, cluster) in [
-            (&live_col, XtreamCluster::Live),
-            (&vod_col, XtreamCluster::Video),
-            (&series_col, XtreamCluster::Series)] {
-            match write_to_file_width_idx(&path, data, &cluster) {
-                Ok(()) => {}
-                Err(err) => {
-                    errors.push(format!("Persisting collection failed: {}: {}", cluster, err));
                 }
             }
         }
@@ -315,14 +307,19 @@ pub(crate) fn xtream_get_all(cfg: &Config, target_name: &str, collection_name: &
     Err(Error::new(std::io::ErrorKind::Other, format!("Cant find collection: {}/{}", target_name, collection_name)))
 }
 
-fn load_map(path: &Path) -> Option<BTreeMap<i32, (u32, u16)>> {
-    match std::fs::read(path) {
+fn load_index(path: &Path) -> Option<IndexTree> {
+    match fs::read(path) {
         Ok(encoded) => {
-            let decoded: BTreeMap<i32, (u32, u16)> = bincode::deserialize(&encoded[..]).unwrap();
+            let decoded: IndexTree = bincode::deserialize(&encoded[..]).unwrap();
             Some(decoded)
         }
         Err(_) => None,
     }
+}
+
+fn write_index(path: &PathBuf, index: &IndexTree) -> std::io::Result<()> {
+    let encoded = bincode::serialize(index).unwrap();
+    fs::write(path, encoded)
 }
 
 fn seek_read(
@@ -337,15 +334,68 @@ fn seek_read(
     Ok(buf)
 }
 
-fn xtream_get_stream_info(cfg: &Config, target_name: &str, stream_id: i32, cluster: XtreamCluster) -> Result<String, Error> {
-    if let Some(path) = get_xtream_storage_path(cfg, target_name) {
-        let (col_path, idx_path) = get_collection_and_idx_path(&path, &cluster);
-        if idx_path.exists() && col_path.exists() {
-            if let Some(idx_map) = load_map(&idx_path) {
-                if let Some((offset, size)) = idx_map.get(&stream_id) {
-                    let mut reader = BufReader::new(File::open(&col_path).unwrap());
-                    if let Ok(bytes) = seek_read(&mut reader, *offset, *size) {
-                        return Ok(String::from_utf8(bytes).unwrap());
+pub(crate) async fn xtream_get_stream_info(app_state: &AppState, target_name: &str, stream_id: i32,
+                                           cluster: XtreamCluster,
+                                           user: &UserCredentials) -> Result<String, Error> {
+    let target_input = app_state.config.get_xtream_input_for_target(target_name);
+    let cache_info = target_input.and_then(|i| i.options.as_ref())
+        .map(|o| o.xtream_info_cache).unwrap_or(false);
+    let mut index_tree: Option<IndexTree> = None;
+    if cache_info {
+        if let Some(path) = get_xtream_storage_path(&app_state.config, target_name) {
+            let (col_path, idx_path) = get_info_collection_and_idx_path(&path, &cluster);
+            let lock = app_state.shared_locks.get_lock(target_name);
+            let shared_lock = lock.read().unwrap();
+            if idx_path.exists() && col_path.exists() {
+                index_tree = load_index(&idx_path);
+                if let Some(idx_map) = &index_tree {
+                    if let Some((offset, size)) = idx_map.get(&stream_id) {
+                        let mut reader = BufReader::new(File::open(&col_path).unwrap());
+                        if let Ok(bytes) = seek_read(&mut reader, *offset, *size) {
+                            let mut decomp: Vec<u8> = Vec::new();
+                            let _ = lzma_rs::lzma_decompress(&mut bytes.as_slice(), &mut decomp);
+                            drop(shared_lock);
+                            return Ok(String::from_utf8(decomp).unwrap());
+                        }
+                    }
+                }
+            }
+            drop(shared_lock);
+        }
+    }
+
+    let (action, stream_id_field) = match cluster {
+        XtreamCluster::Live => ("get_live_info", "live_id"),
+        XtreamCluster::Video => ("get_vod_info", "vod_id"),
+        XtreamCluster::Series => ("get_series_info", "series_id"),
+    };
+
+    // not indexed, receive
+    match target_input {
+        None => {}
+        Some(input) => {
+            let info_url = format!("{}/player_api.php?username={}&password={}&action={}&{}={}",
+                                   input.url, &user.username, &user.password, action,
+                                   stream_id_field, stream_id);
+            let url = reqwest::Url::parse(&info_url).unwrap();
+            let client = get_client_request(input, url);
+            if let Ok(response) = client.send().await {
+                if response.status().is_success() {
+                    match response.text().await {
+                        Ok(content) => {
+                            if cache_info {
+                                if index_tree.is_none() {
+                                    index_tree = Some(IndexTree::new());
+                                }
+                                match write_xtream_info(app_state, target_name, stream_id, &cluster, &content,
+                                                        index_tree.as_mut().unwrap()) {
+                                    Ok(_) => {}
+                                    Err(err) => { error!("{}", err.to_string()); }
+                                }
+                            }
+                            return Ok(content);
+                        }
+                        Err(err) => { error!("Failed to download info {}", err.to_string()); }
                     }
                 }
             }
@@ -354,93 +404,28 @@ fn xtream_get_stream_info(cfg: &Config, target_name: &str, stream_id: i32, clust
     Err(Error::new(std::io::ErrorKind::Other, format!("Cant find stream with id: {}/{}/{}", target_name, &cluster, stream_id)))
 }
 
-pub(crate) fn xtream_get_series_info(cfg: &Config, target_name: &str, stream_id: i32) -> Result<String, Error> {
-    /*
-    {
-        "episodes": {
-        "": [
-        {
-            "added": string,
-            "container_extension": string,
-            "custom_sid": string,
-            "direct_source": string,
-            "episode_num": int,
-            "id": string,
-            "info": {
-                "bitrate": int,
-                "duration": string,
-                "duration_secs": int,
-                "movie_image": string,
-                "name": string,
-                "plot": string,
-                "rating": float,
-                "releasedate": string,
-                "audio": FFMPEGStreamInfo,
-                "video": FFMPEGStreamInfo
+pub(crate) async fn xtream_get_short_epg(app_state: &AppState, target_name: &str, stream_id: &str, limit: &str, user: &UserCredentials) -> Result<String, Error> {
+    match app_state.config.get_xtream_input_for_target(target_name) {
+        None => {}
+        Some(input) => {
+            let mut info_url = format!("{}/player_api.php?username={}&password={}&action=get_short_epg&stream_id={}",
+                                       input.url, &user.username, &user.password, stream_id);
+            if !(limit.is_empty() || limit.eq("0")) {
+                info_url = format!("{}&limit={}", info_url, limit);
             }
-            "season": int,
-            "title": string
-         }
-        ]
-    },
-        "info": {
-            "backdrop_path: [string],
-            "cast":  string,
-            "category_id": string,
-            "cover":  string,
-            "director":  string,
-            "episode_run_time":  string,
-            "genre":  string,
-            "last_modified": string,
-            "name":  string,
-            "num":  int,
-            "plot":  string,
-            "rating, string,
-            "rating_5based": float,
-            "releaseDate": string,
-            "series_id": int,
-            "stream_type": string,
-            "youtube_trailer": string,
+            let url = reqwest::Url::parse(&info_url).unwrap();
+            let client = get_client_request(input, url);
+            if let Ok(response) = client.send().await {
+                if response.status().is_success() {
+                    match response.text().await {
+                        Ok(content) => {
+                            return Ok(content);
+                        }
+                        Err(err) => { error!("Failed to download epg {}", err.to_string()); }
+                    }
+                }
+            }
+        }
     }
-    }
-    "seasons": []
-}
-        */
-    // TODO restructure
-    xtream_get_stream_info(cfg, target_name, stream_id, XtreamCluster::Series)
-}
-
-pub(crate) fn xtream_get_vod_info(cfg: &Config, target_name: &str, stream_id: i32) -> Result<String, Error> {
-    /*
-    {
-    "info": {
-        "backdrop_path": [string],
-        "bitrate": FlexInt,
-        "cast": string,
-        "director": string,
-        "duration": string,
-        "duration_secs": FlexInt,
-        "genre": string,
-        "movie_image": string,
-        "plot": string,
-        "rating": FlexFloat,
-        "releasedate": string,
-        "tmdb_id": int,
-        "youtube_trailer": string,
-        "audio": FFMPEGStreamInfo,
-        "video": FFMPEGStreamInfo,
-	} `json:"info"`
-	"movie_data": {
-		"added": string,
-		"category_id": string,
-		"container_extension": string,
-		"custom_sid": string,
-		"direct_source": string,
-		"name": string,
-		"stream_id": int
-	}
-	}
-     */
-  // TODO restructure
-    xtream_get_stream_info(cfg, target_name, stream_id, XtreamCluster::Video)
+    Err(Error::new(std::io::ErrorKind::Other, format!("Cant find short epg with id: {}/{}", target_name, stream_id)))
 }